+++ CHANGELOG.md
@@ -150,6 +150,7 @@
 - New #1057: Add `CompositeExpression` class (@vjik)
 - Chg #1058: Refactor `Expression` class: declare `$expression` and `$params` as public readonly properties, remove 
   `getParams()` method (@vjik)
+- New #1062: `Stringable` value support in `Like` and `NotLike` conditions (@vjik)
 
 ## 1.3.0 March 21, 2024